+++ backend/src/main/java/backend/mulkkam/intake/dto/DateRangeRequest.java
@@ -1,28 +1,13 @@
 package backend.mulkkam.intake.dto;
 
-import static backend.mulkkam.common.exception.errorCode.BadRequestErrorCode.INVALID_DATE_RANGE;
-
-import backend.mulkkam.common.exception.CommonException;
+import backend.mulkkam.intake.service.vo.DateRange;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 
 public record DateRangeRequest(
         LocalDate from,
         LocalDate to
 ) {
-
-    public DateRangeRequest {
-        if (from.isAfter(to)) {
-            throw new CommonException(INVALID_DATE_RANGE);
-        }
-    }
-
-    public LocalDateTime startDateTime() {
-        return from.atStartOfDay();
-    }
-
-    public LocalDateTime endDateTime() {
-        return to.atTime(LocalTime.MAX);
+    public DateRange toDateRange() {
+        return new DateRange(from, to);
     }
 }
